feat(gtag): add gtag and create files for different function
