+++ .prettierignore
@@ -1,3 +1,3 @@
 **/.temp
-**/.cache
+**/cache
 **/dist
