--- docs/.vitepress/client.ts
@@ -1,11 +1,0 @@
-import { defineClientConfig } from '@vuepress/client'
-import Layout from './layouts/Layout.vue'
-
-export default defineClientConfig({
-  enhance({ app, router, siteData }) {},
-  setup() {},
-  rootComponents: [],
-  layouts: {
-    Layout,
-  },
-})
