+++ docs/.vitepress/config.ts
@@ -1,88 +1,11 @@
-import { DefaultTheme, defineConfig } from 'vitepress'
-
-import glob from 'glob'
-
+import { defineConfig } from 'vitepress'
 import * as dotenv from 'dotenv'
 
+import gtagHead from './typescript/node/generateSidebar'
+import generateSidebar from './typescript/node/sidebat'
+import generateRewrites from './typescript/node/generateRewrites'
+
 dotenv.config()
-
-const sidebar: DefaultTheme.Sidebar = []
-
-glob
-  .sync('docs/**/*.md')
-  .sort((a, b) => {
-    const aArray = a.split('/')
-    const bArray = b.split('/')
-
-    for (let i = 0; i < Math.min(aArray.length, bArray.length); i++) {
-      const aName = aArray[i]
-      const bName = bArray[i]
-      if (aName !== bName) {
-        if (aName === 'index.md') {
-          return -1
-        }
-        if (bName === 'index.md') {
-          return 1
-        }
-        if (aName.endsWith('.md') && !bName.endsWith('.md')) {
-          return -1
-        }
-        if (!aName.endsWith('.md') && bName.endsWith('.md')) {
-          return 1
-        }
-        return aName.localeCompare(bName)
-      }
-    }
-
-    return 0
-  })
-  .map((path) => path.replace('docs/', '').replace(/_/g, ''))
-  .forEach((path) =>
-    path.split('/').forEach((text, index, array) => {
-      let items = sidebar
-
-      for (let i = 0; i < index; i++) {
-        items = (
-          items.find(
-            (child) => typeof child === 'object' && child.text === array[i]
-          ) as DefaultTheme.SidebarMulti
-        ).items
-      }
-
-      if (text === 'index.md') {
-        const pageName = array[index - 1]
-
-        items.push({
-          text: pageName ? `${pageName}介紹` : '首頁',
-          link: `/${path.replace('index.md', '')}`,
-        })
-        return
-      }
-
-      if (text.endsWith('.md')) {
-        items.push({
-          text: text.replace('.md', ''),
-          link: `/${path.replace('.md', '')}`,
-        })
-        return
-      }
-
-      const child = items.find(
-        (child) => typeof child === 'object' && child.text === text
-      ) as DefaultTheme.Sidebar
-
-      if (!child) {
-        items.push({ text: text, items: [], collapsed: true })
-      }
-    })
-  )
-
-const rewrites = Object.fromEntries(
-  glob
-    .sync('docs/**/*.md')
-    .map((path) => path.replace('docs/', ''))
-    .map((path) => [path, path.replace(/_/g, '')])
-)
 
 export default defineConfig({
   base: '/',
@@ -96,9 +19,10 @@ export default defineConfig({
         content: <string>process.env.VITE_APP_GOOGLE_META_TAG_CONTENT,
       },
     ],
+    ...gtagHead,
   ],
   themeConfig: {
-    sidebar,
+    sidebar: generateSidebar(),
     editLink: {
       pattern: ({ relativePath }) => {
         return `https://github.com/travel-guide-tw/travel-guide-tw.github.io/edit/main/docs/${relativePath}`
@@ -107,5 +31,5 @@ export default defineConfig({
     },
   },
   title: '台灣軟體人旅遊指南',
-  rewrites,
+  rewrites: generateRewrites(),
 })
